requirements.txt
Flask==2.3.2
gunicorn==21.2.0
Pillow==9.5.0
google-cloud-storage==2.9.0
python-dotenv==1.0.0